feat(#5): 날씨 데이터 컴포넌트화
- 날씨 데이터를 다른 컴포넌트에서 활용할 수 있도록 수정
- 도시 이름에 스타일 적용
